# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TransitOps API",
    "description": "Fleet operations platform \u2014 manage vehicles, drivers, trips, maintenance, fuel, expenses, and reports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/reports/fuel-efficiency": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Fuel efficiency report",
        "description": "Distance / Fuel per vehicle. Fuel data sourced from FuelLog (canonical).",
        "operationId": "fuel_efficiency_report_reports_fuel_efficiency_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FuelEfficiencyReport"
                  },
                  "type": "array",
                  "title": "Response Fuel Efficiency Report Reports Fuel Efficiency Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/fuel-efficiency/csv": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Fuel efficiency report (CSV)",
        "operationId": "fuel_efficiency_csv_reports_fuel_efficiency_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/fleet-utilization": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Fleet utilization report",
        "description": "Vehicles On Trip / total active vehicles.",
        "operationId": "fleet_utilization_report_reports_fleet_utilization_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetUtilizationReport"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/fleet-utilization/csv": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Fleet utilization report (CSV)",
        "operationId": "fleet_utilization_csv_reports_fleet_utilization_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/operational-cost": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Operational cost report",
        "description": "Fuel + Maintenance + Expenses per vehicle.",
        "operationId": "operational_cost_report_reports_operational_cost_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperationalCostReport"
                  },
                  "type": "array",
                  "title": "Response Operational Cost Report Reports Operational Cost Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/operational-cost/csv": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Operational cost report (CSV)",
        "operationId": "operational_cost_csv_reports_operational_cost_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/vehicle-roi": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Vehicle ROI report",
        "description": "(Revenue \u2212 (Maintenance + Fuel)) / Acquisition Cost per vehicle.",
        "operationId": "vehicle_roi_report_reports_vehicle_roi_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehicleROIReport"
                  },
                  "type": "array",
                  "title": "Response Vehicle Roi Report Reports Vehicle Roi Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/reports/vehicle-roi/csv": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Vehicle ROI report (CSV)",
        "operationId": "vehicle_roi_csv_reports_vehicle_roi_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FleetUtilizationReport": {
        "properties": {
          "total_active_vehicles": {
            "type": "integer",
            "title": "Total Active Vehicles"
          },
          "vehicles_on_trip": {
            "type": "integer",
            "title": "Vehicles On Trip"
          },
          "utilization_pct": {
            "type": "number",
            "title": "Utilization Pct"
          }
        },
        "type": "object",
        "required": [
          "total_active_vehicles",
          "vehicles_on_trip",
          "utilization_pct"
        ],
        "title": "FleetUtilizationReport"
      },
      "FuelEfficiencyReport": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "total_distance_km": {
            "type": "number",
            "title": "Total Distance Km"
          },
          "total_fuel_liters": {
            "type": "number",
            "title": "Total Fuel Liters"
          },
          "efficiency_km_per_liter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Efficiency Km Per Liter"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "registration_number",
          "total_distance_km",
          "total_fuel_liters",
          "efficiency_km_per_liter"
        ],
        "title": "FuelEfficiencyReport"
      },
...
      "OperationalCostReport": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "fuel_cost": {
            "type": "number",
            "title": "Fuel Cost"
          },
          "maintenance_cost": {
            "type": "number",
            "title": "Maintenance Cost"
          },
          "expense_cost": {
            "type": "number",
            "title": "Expense Cost"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "registration_number",
          "fuel_cost",
          "maintenance_cost",
          "expense_cost",
          "total_cost"
        ],
        "title": "OperationalCostReport"
      },
...
      "VehicleROIReport": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "registration_number": {
            "type": "string",
            "title": "Registration Number"
          },
          "total_revenue": {
            "type": "number",
            "title": "Total Revenue"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "acquisition_cost": {
            "type": "number",
            "title": "Acquisition Cost"
          },
          "roi_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roi Pct"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "registration_number",
          "total_revenue",
          "total_cost",
          "acquisition_cost",
          "roi_pct"
        ],
        "title": "VehicleROIReport"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}